+++ gevent/hub.py
@@ -98,11 +98,15 @@ def idle(priority=0):
 
 
 def kill(greenlet, exception=GreenletExit):
-    """Kill greenlet asynchronously. The current greenlet is not unscheduled.
+    """
+    Kill greenlet asynchronously. The current greenlet is not unscheduled.
 
-    Note, that :meth:`gevent.Greenlet.kill` method does the same and more. However,
-    MAIN greenlet - the one that exists initially - does not have ``kill()`` method
-    so you have to use this function.
+    .. note::
+
+        The method :meth:`gevent.Greenlet.kill` method does the same and
+        more (and the same caveats listed there apply here). However, the MAIN
+        greenlet - the one that exists initially - does not have a
+        ``kill()`` method so you have to use this function.
     """
     if not greenlet.dead:
         get_hub().loop.run_callback(greenlet.throw, exception)
@@ -150,9 +154,34 @@ class signal(object):
 
 
 def reinit():
+    # An internal, undocumented function. Called by gevent.os.fork
+    # in the child process. The loop reinit function in turn calls
+    # libev's ev_loop_fork function.
     hub = _get_hub()
     if hub is not None:
         hub.loop.reinit()
+        # libev's fork watchers are slow to fire because the only fire
+        # at the beginning of a loop; due to our use of callbacks that
+        # run at the end of the loop, that may be too late. The
+        # threadpool and resolvers depend on the fork handlers being
+        # run ( specifically, the threadpool will fail in the forked
+        # child if there were any threads in it, which there will be
+        # if the resolver_thread was in use (the default) before the
+        # fork.)
+        #
+        # If the forked process wants to use the threadpool or
+        # resolver immediately, it would hang.
+        #
+        # The below is a workaround. Fortunately, both of these
+        # methods are idempotent and can be called multiple times
+        # following a fork if the suddenly started working, or were
+        # already working on some platforms. Other threadpools and fork handlers
+        # will be called at an arbitrary time later ('soon')
+        if hasattr(hub.threadpool, '_on_fork'):
+            hub.threadpool._on_fork()
+        # resolver_ares also has a fork watcher that's not firing
+        if hasattr(hub.resolver, '_on_fork'):
+            hub.resolver._on_fork()
 
 
 def get_hub_class():
@@ -637,7 +666,7 @@ def iwait(objects, timeout=None, count=None):
     """
     Yield objects as they are ready, until all (or `count`) are ready or `timeout` expired.
 
-    :param objects: A list (supporting `len`) containing objects
+    :param objects: A sequence (supporting :func:`len`) containing objects
         implementing the wait protocol (rawlink() and unlink()).
     :param count: If not `None`, then a number specifying the maximum number
         of objects to wait for.
@@ -696,7 +725,7 @@ def wait(objects=None, timeout=None, count=None):
     - all servers were stopped
     - all event loop watchers were stopped.
 
-    If ``count`` is ``None`` (the default), wait for all ``object``s
+    If ``count`` is ``None`` (the default), wait for all ``objects``
     to become ready.
 
     If ``count`` is a number, wait for (up to) ``count`` objects to become