+++ changelog.rst
@@ -45,6 +45,19 @@ Unreleased
 - ``gevent.iwait`` no longer throws ``LoopExit`` if the caller
   switches greenlets between return values. Reported and initial patch
   in :pr:`467` by Alexey Borzenkov.
+- The default threadpool and default threaded resolver work in a
+  forked child process, such as with ``multiprocessing.Process``.
+  Previously the child process would hang indefinitely. Reported in
+  :issue:`230` by Lx Yu.
+- Fork watchers are more likely to (eventually) get called in a
+  multi-threaded program.
+- ``gevent.killall`` accepts an arbitrary iterable for the greenlets
+  to kill. Reported in :issue:`404` by Martin Bachwerk; seen in
+  combination with older versions of simple-requests.
+- ``gevent.local.local`` objects are now eligible for garbage
+  collection as soon as the greenlet finishes running, matching the
+  behaviour of the built-in ``threading.local`` (when implemented
+  natively). Reported in :issue:`387` by AusIV.
 
 1.1a1 (Jun 29, 2015)
 ====================